pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ds-poc-template"
version = "0.1.0"
description = "Template liviano para proyectos DS/MLE en PoC"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name="iData", email="data@idata.local"}]
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "pyyaml>=6.0",
  "scikit-learn>=1.3",
  "joblib>=1.3",
  "typer>=0.12",
  "rich>=13.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.6",
  "black>=24.0",
]

[project.scripts]
project = "project_name.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E","F","I","B","UP","N","S","W"]
ignore = ["S101"] # allow assert in tests
fix = true
